% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{bem_new}
\alias{bem_new}
\title{Create a new object with BEM data}
\usage{
bem_new(
  CP,
  CA,
  CB,
  CTM,
  CTO,
  CQ,
  CTL,
  CK1,
  CK4,
  ACT,
  RA,
  RB,
  RTM,
  RTO,
  RQ,
  FA,
  UA = 0.1,
  SDA = 0.15,
  ED = 3000
)
}
\arguments{
\item{CP}{proportion of maximum consumption}

\item{CA}{intercept of allometric mass function}

\item{CB}{slope of allometric mass function; should be negative bc bigger fish consume less per gram of body mass than smaller fish}

\item{CTM}{critical thermal maximum (degrees C)}

\item{CTO}{laboratory temperature preferendum (degrees C)}

\item{CQ}{approximates a Q10; the rate at which the function increases over relatively low water temperatures}

\item{CTL}{temperature at which consumption is XXXX (degrees C)}

\item{CK1}{temperature at which consumption is XXXX (degrees C)}

\item{CK4}{temperature at which consumption is XXXX (degrees C)}

\item{ACT}{activity coefficient}

\item{RA}{intercept of allometric activity function}

\item{RB}{slope of allometric activity function}

\item{RTM}{critical thermal maximum (degrees C)}

\item{RTO}{laboratory temperature preferendum (degrees C)}

\item{RQ}{approximates a Q10; the rate at which the function increases over relatively low water temperatures}

\item{FA}{fish body mass (grams)}

\item{UA}{activity (mg/kg-day)}

\item{SDA}{specific dynamic action (proportion)}

\item{ED}{energy density (J/kg)}
}
\value{
BEM object
}
\description{
Create a new object with BEM data
}
