% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{bem_sensitive}
\alias{bem_sensitive}
\title{Perform sensitivity analysis to determine relative effects of input parameters on simulated growth}
\usage{
bem_sensitive(
  start_M2 = 100,
  temperature,
  parms.intrinsic,
  parms.temporal,
  C_eq,
  R_eq,
  parm.dev
)
}
\arguments{
\item{start_M2}{(default is 100 grams).}

\item{temperature}{a dataframe populated with a time series of mean daily water temperature (degrees C x 10) of a habitat patch.}

\item{parms.intrinsic}{A parms.intrinsic object; note that temperature dependent parameters are bypassed if C_eq = 4 and/or R_eq = 4.}

\item{parms.temporal}{a dataframe populated with a time series of intrinsic (e.g., GSI) and extrinsic (e.g., prey energy density) biological parameters.}

\item{C_eq}{Specify consumption equation 1, 2, or 3 from Hanson et al. 1997 or equation 4.}

\item{R_eq}{Specify respiration equation 1 or 2 from Hanson et al. 1997.}

\item{parm_dev}{Specify specify the percent deviation for each intrinsic parameter.}
}
\value{
a list of length 2; table of parameter sensitivities; plotted parameter sensitivies.
}
\description{
quantify effect of intrinsic parameter variation on simulated mass at the end of a growth interval from Bartell et al. (1986) CJFAS 43.1.
}
