% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{consumption}
\alias{consumption}
\title{Consumption of energy}
\usage{
consumption(T, W, BEM)
}
\arguments{
\item{T}{temperature (degrees C) at which consumption is calculated}

\item{W}{weight (grams) of fish}

\item{BEM}{object}
}
\value{
specific consumption rate (grams of food consumed per gram of fish mass per day)
}
\description{
Consumption of energy
}
