% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{consumption4}
\alias{consumption4}
\title{Consumption equation 4}
\usage{
consumption4(M, T, CP = 1, parms.intrinsic, match.table)
}
\arguments{
\item{M}{Mass (grams) of fish.}

\item{T}{Temperature (degrees C) at which consumption is calculated.}

\item{CP}{Proportion of maximum consumption (default is 1.0).}

\item{parms.intrinsic}{A parms.intrinsic object; note that temperature dependent parameters are bypassed.}

\item{match.table}{temperature dependent parameters formatted as a table with three columns named WT_mean, C_match, and R_match.}
}
\value{
Specific consumption rate (grams of food consumed per gram of fish mass per day).
}
\description{
Consumption equation 4; new equation for fishyenergy that simply matches Cmax for a given temperature.
}
