% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{daily_temperature_raw}
\alias{daily_temperature_raw}
\title{Get daily temperature data from monitoring stations}
\usage{
daily_temperature_raw(site_code, start_date, end_date)
}
\arguments{
\item{site_code}{WQP site code (e.g. "USGS-12345678")}

\item{start_date}{Start date for the data retrieval (e.g. "2020-01-01")}

\item{end_date}{End date for the data retrieval (e.g. "2020-12-31")}
}
\value{
A data frame with daily temperature data
}
\description{
This uses the dataRetrieval package to fetch daily temperature data
}
