% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\docType{data}
\name{envir_casc}
\alias{envir_casc}
\title{Environmental covariates for many stream reaches}
\format{
A vector shapefile with an attribute table containing 513 rows and 5 columns:
\describe{
  \item{COMID}{(character) Common Identifier for NHD reaches.}
  \item{GNIS_NAME}{(character) Stream name from NHD after Geographic Names Information System.}
  \item{FCode}{(integer) Flowline type from NHD.}
  \item{TotDASqKM}{(numeric) Catchment area draining to the interconfluence stream reach in square kilometers.}
  \item{ElevCat}{(numeric) Median elevation of the interconfluence stream reach in meters above sea level.}
  ...
}
}
\source{
Southeastern Climate Adaptation Science Center (CASC), National Hydrography Dataset (NHD), StreamCat
}
\usage{
envir_casc
}
\description{
This dataset contains geographic and hydrographic covariates for all 513 interconfluence stream reaches in the Little River, TN.
}
\keyword{datasets}
