% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\docType{data}
\name{envir_lake}
\alias{envir_lake}
\title{Daily water temperature many lakes}
\format{
A data frame with 365 rows and 8046 columns:
\describe{
\item{Hylak_id}{(numeric) mean daily water temperature in degrees C for all 8046 unique identifier of lake, from HydroLAKES (Messager et al., 2016)}
\item{center_long}{(numeric) longitude of lake center in decimal degrees}
\item{center_lat}{(numeric) latitude of lake center in decimal degrees}
\item{stat_method}{(numeric) see Korver et al. 2024 for description}
\item{n_obs}{(numeric) see Korver et al. 2024 for description}
\item{intermittency}{(numeric) see Korver et al. 2024 for description}
\item{Lake_type}{(numeric) 1 = natural lake, 2 = reservoir, 3 = natural lake with regulation structure}
\item{Elevation}{(numeric) lake surface elevation in meters above sea level}
\item{Lake_area}{(numeric) lake surface area in square kilometers}
\item{Depth_avg}{(numeric) mean lake depth in meters}
...
}
}
\source{
LakeTEMP; Korver et al. 2024. Remote Sensing of Environment. https://doi.org/10.1016/j.rse.2024.114164.
}
\usage{
envir_lake
}
\description{
This dataset contains modeled maximum daily water temperature for every day of the year 2022 for all interconfluence reaches of the Little River, TN.
}
\keyword{datasets}
