% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\docType{data}
\name{fb4_data}
\alias{fb4_data}
\title{Parameters for Bioenergetics Model (BEM) from FB4}
\format{
An object of class \code{data.frame} with 105 rows and 44 columns.
}
\source{
: Deslauriers et al. (1997), https://doi.org/10.1080/03632415.2017.1377558 and https://github.com/jim-breck/FB4/blob/master/Parameters_official.csv, as well as the original source for each row of data.
}
\usage{
fb4_data
}
\description{
This dataset comes from the Fish Bioenergetics 4 (FB4) model, version 1.17, and is from https://github.com/jim-breck/FB4/blob/master/Parameters_official.csv. If you use this, please cite the original source (see fb4_data$Source) as well as Deslauriers et al. (1997), https://doi.org/10.1080/03632415.2017.1377558. It contains common names for aquatic species, scientific names and taxonomy, the life stage modeled, source, and parameters for the BEM.

@format A data frame with 105 rows and 44 variables
}
\keyword{datasets}
