% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{Find_Closest_FB4_Species}
\alias{Find_Closest_FB4_Species}
\title{Find closest relatives of a species in the FB4 dataset}
\usage{
Find_Closest_FB4_Species(species)
}
\arguments{
\item{species}{Scientific name of the species (e.g. "Micropterus salmoides")}
}
\value{
A vector of scientific names of the closest relatives in the FB4 dataset
}
\description{
This uses the Open Tree of Life to find relatives of a species.
}
