% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{lifestages_from_fb4}
\alias{lifestages_from_fb4}
\title{Get available life stages for a species in the FB4 dataset}
\usage{
lifestages_from_fb4(species)
}
\arguments{
\item{species}{Scientific name of the species (e.g. "Micropterus salmoides")}
}
\value{
A vector of unique life stages for the species
}
\description{
Get available life stages for a species in the FB4 dataset
}
