% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{loop_species_across_station}
\alias{loop_species_across_station}
\title{Loop through species across a station}
\usage{
loop_species_across_station(
  T_vector,
  starting_weight = 6.382417,
  prey_ED = 3698,
  oxycal_coeff = 13560
)
}
\arguments{
\item{T_vector}{Temperatures (degrees C), starting at Julian day 1}

\item{starting_weight}{Starting weight (grams) for the species; can be a single value or a vector with one value per organism}

\item{prey_ED}{joules per gram of wet mass for the species' prey; can be a single value or a vector with one value per organism}

\item{oxycal_coeff}{joules per gram of oxygen for the species; can be a single value or a vector with one value per organism}

\item{species}{Scientific name of the species (e.g. "Micropterus salmoides")}
}
\value{
A data.frame with growth results for each species and life stage, including species name, life stage, and cumulative weight over the year
}
\description{
ADD LATER
}
