% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{new_BEM}
\alias{new_BEM}
\title{Create a new object with BEM data}
\usage{
new_BEM(CP, CA, CB, CTM, CTO, CQ, ACT, RA, RB, RTM, RTO, RQ, FA, UA, SDA, ED)
}
\arguments{
\item{CP}{proportion of maximum consumption}

\item{CA}{intercept of allometric mass function}

\item{CB}{slope of allometric mass function; should be negative bc bigger fish consume less per gram of body mass than smaller fish}

\item{CTM}{critical thermal maximum (degrees C)}

\item{CTO}{laboratory temperature preferendum (degrees C)=}

\item{CQ}{approximates a Q10; the rate at which the function increases over relatively low water temperatures}

\item{ACT}{activity coefficient}

\item{RA}{intercept of allometric activity function}

\item{RB}{slope of allometric activity function}

\item{RTM}{critical thermal maximum (degrees C)}

\item{RTO}{laboratory temperature preferendum (degrees C)}

\item{RQ}{approximates a Q10; the rate at which the function increases over relatively low water temperatures}

\item{FA}{fish body mass (grams)}

\item{UA}{activity (mg/kg-day)}

\item{SDA}{salinity (psu)}

\item{ED}{energy density (J/kg)}
}
\value{
BEM object
}
\description{
Create a new object with BEM data
}
