% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{respiration2}
\alias{respiration2}
\title{Respiration equation 2}
\usage{
respiration2(M, T, ACT = 1, parms.intrinsic)
}
\arguments{
\item{M}{Mass (grams) of fish.}

\item{T}{Temperature (degrees C) at which respiration is calculated.}

\item{ACT}{Activity multiplier (default is 1.0).}

\item{parms.intrinsic}{A parms.intrinsic object.}
}
\value{
Specific respiration rate (grams of oxygen per gram of fish mass per day).
}
\description{
Respiration equation 2 from Hanson et al. 1997.
}
