% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{respiration4}
\alias{respiration4}
\title{Respiration equation 4}
\usage{
respiration4(M, T, ACT = 1, parms.intrinsic, match.table)
}
\arguments{
\item{M}{Mass (grams) of fish}

\item{T}{Temperature (degrees C) at which consumption is calculated}

\item{ACT}{Activity multiplier (default is 1.0)}

\item{parms.intrinsic}{A parms.intrinsic object; note that temperature dependent parameters are bypassed}

\item{match.table}{temperature dependent parameters formatted as a table with three columns named WT_mean, C_match, and R_match}
}
\value{
Specific respiration rate (grams of oxygen per gram of fish mass per day)
}
\description{
Respiration equation 4; new equation for fishyenergy that simply matches Rrest for a given temperature
}
