% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{single_station_compute}
\alias{single_station_compute}
\title{Compute growth for a year at a station}
\usage{
single_station_compute(
  T_vector,
  BEM,
  starting_weight = 6.382417,
  prey_ED = 3698,
  oxycal_coeff = 13560
)
}
\arguments{
\item{T_vector}{Temperatures (degrees C), starting at Julian day 1}

\item{BEM}{BEM object}

\item{starting_weight}{Starting weight (grams) from Groeschel-Taylor et al}

\item{prey_ED}{joules per gram of wet mass}

\item{oxycal_coeff}{joules per gram of oxygen}
}
\value{
data.frame with growth over year
}
\description{
ADD LATER
}
