% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\docType{data}
\name{temps_lake}
\alias{temps_lake}
\title{Daily water temperature for many lakes}
\format{
A data frame with 365 rows and 8,046 columns:
\describe{
  \item{date}{(date) dates from 1 January 2021 to 31 December 2021; note that temps are the mean from years 2013 to 2021}
  \item{Hylak_id}{(numeric) mean daily water temperature in degrees C for all 8046 unique identifier of lake, from HydroLAKES (Messager et al., 2016)}
  ...
}
}
\source{
LakeTEMP; Korver et al. 2024. Remote Sensing of Environment. https://doi.org/10.1016/j.rse.2024.114164.
}
\usage{
temps_lake
}
\description{
This dataset contains modeled maximum daily water temperature for every day of the year 2022 for 8,046 lakes in 15 southeastern United States plus Puerto Rico.
}
\keyword{datasets}
