% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{bem_from_fb4}
\alias{bem_from_fb4}
\title{Create a BEM object from FB4 data}
\usage{
bem_from_fb4(
  species,
  lifestage = NULL,
  CP = 0.5,
  FA = 0.1018428,
  UA = 0.08743042,
  SDA = 0.150819,
  ED = 3983.646
)
}
\arguments{
\item{species}{Scientific name of the species (e.g. "Micropterus salmoides")}

\item{lifestage}{Life stage of the species (e.g. "Adult"). If NA, uses the first life stage found for the species.}

\item{CP}{proportion of maximum consumption (default is 0.5)}

\item{FA}{fish body mass (grams) (default is 0.1018428)}

\item{UA}{activity (mg/kg-day) (default is 0.08743042)}

\item{SDA}{salinity (psu) (default is 0.150819)}

\item{ED}{energy density (J/kg) (default is 3983.646)}
}
\value{
BEM object with parameters for the specified species and life stage
}
\description{
Create a BEM object from FB4 data
}
