% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{bem_project}
\alias{bem_project}
\title{Project growth and performance indices across heterogeneous habitat patches}
\usage{
bem_project(
  start_M2 = 100,
  temperature,
  parms.intrinsic,
  parms.temporal,
  C_eq,
  R_eq
)
}
\arguments{
\item{start_M2}{(default is 100 grams)}

\item{temperature}{a list of dataframes populated with a time series of mean daily water temperature (degrees C); each list element is a different habitat patch (default is 20 degrees C)}

\item{parms.intrinsic}{A parms.intrinsic object}

\item{parms.temporal}{a list of dataframes populated with a time series of intrinsic (e.g., GSI) and extrinsic (e.g., prey energy density) biological parameters; each list element is a different habitat patch}

\item{C_eq}{Specify consumption equation 1, 2, or 3 from Hanson et al. 1997}

\item{R_eq}{Specify respiration equation 1 or 2 from Hanson et al. 1997}
}
\value{
a list of length 2; raw bem_grow output for each patch and year-end performance indices for each patch
}
\description{
solve daily energy budgets for n habitat patches
}
