% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{bem_validate}
\alias{bem_validate}
\title{Project growth to validate a bioenergetics model with empirical end-of-year mass}
\usage{
bem_validate(
  start_M2 = 100,
  end_M2.empirical,
  temperature = 20,
  parms.intrinsic,
  parms.temporal,
  C_eq,
  R_eq,
  resamp.n = 1000
)
}
\arguments{
\item{start_M2}{(default is 100 grams)}

\item{end_M2.empirical}{mass of fish at end of year based on empirically-derived field or lab size-at-age estimate}

\item{temperature}{a dataframe populated with a time series of mean daily water temperature (degrees C) of a habitat patch (default is 20 degrees C)}

\item{parms.intrinsic}{A parms.intrinsic object}

\item{parms.temporal}{a dataframe populated with a time series of intrinsic (e.g., GSI) and extrinsic (e.g., prey energy density) biological parameters}

\item{C_eq}{Specify consumption equation 1, 2, or 3 from Hanson et al. 1997}

\item{R_eq}{Specify respiration equation 1 or 2 from Hanson et al. 1997}

\item{resamp.n}{specify the number of CP-ACT parameter sets to draw from uniform distributions (default is 1000)}
}
\value{
a list of length 5; raw bem_grow output for each patch and year-end performance indices for each patch; also plotted validation output
}
\description{
simulate growth with resampled CP and ACT parameters to identify parameter sets that match empirical end-of-year mass
}
