% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{consumption2}
\alias{consumption2}
\title{Consumption equation 2}
\usage{
consumption2(M, T, CP = 1, parms.intrinsic)
}
\arguments{
\item{M}{mass (grams) of fish.}

\item{T}{temperature (degrees C) at which consumption is calculated.}

\item{CP}{proportion of maximum consumption (default is 1.0).}

\item{parms.intrinsic}{A parms.intrinsic object.}
}
\value{
specific consumption rate (grams of food consumed per gram of fish mass per day).
}
\description{
consumption equation 2 from Hanson et al. 1997.
}
