% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\docType{data}
\name{envir_lake}
\alias{envir_lake}
\title{Environmental covariates for many lakes}
\format{
A data frame with 8,046 rows and 10 columns:
\describe{
  \item{Hylak_id}{(numeric) mean daily water temperature in degrees C for all 8046 unique identifier of lake, from HydroLAKES (Messager et al., 2016)}
  \item{center_long}{(numeric) longitude of lake center in decimal degrees}
  \item{center_lat}{(numeric) latitude of lake center in decimal degrees}
  \item{stat_method}{(numeric) see Korver et al. 2024 for description}
  \item{n_obs}{(numeric) see Korver et al. 2024 for description}
  \item{intermittency}{(numeric) see Korver et al. 2024 for description}
  \item{Lake_type}{(numeric) 1 = natural lake, 2 = reservoir, 3 = natural lake with regulation structure}
  \item{Elevation}{(numeric) lake surface elevation in meters above sea level}
  \item{Lake_area}{(numeric) lake surface area in square kilometers}
  \item{Depth_avg}{(numeric) mean lake depth in meters}
  ...
}
}
\source{
LakeTEMP; Korver et al. 2024. Remote Sensing of Environment. https://doi.org/10.1016/j.rse.2024.114164.
}
\usage{
envir_lake
}
\description{
This dataset contains geographic and bathymetric covariates for all 8,046 lakes extracted from the lakeTEMP dataset.
}
\keyword{datasets}
