% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{respiration}
\alias{respiration}
\title{Respiration 1 equation}
\usage{
<<<<<<< Updated upstream:man/respiration.Rd
respiration(T, W, BEM)
=======
respiration1(M, T, ACT = 1, parms.intrinsic)
>>>>>>> Stashed changes:man/respiration1.Rd
}
\arguments{
\item{M}{Mass (grams) of fish}

\item{T}{Temperature (degrees C) at which respiration is calculated}

\item{ACT}{Activity multiplier (default is 1.0)}

\item{parms.intrinsic}{A parms.intrinsic object}
}
\value{
Specific respiration rate (grams of oxygen per gram of fish mass per day)
}
\description{
<<<<<<< Updated upstream:man/respiration.Rd
Respiration 1 equation
=======
Respiration equation 1 from Hanson et al. 1997
>>>>>>> Stashed changes:man/respiration1.Rd
}
