% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\name{respiration2}
\alias{respiration2}
\title{Respiration equation 2}
\usage{
respiration2(T, W, BEM)
}
\arguments{
\item{T}{temperature (degrees C) at which consumption is calculated}

\item{W}{weight (grams) of fish}

\item{BEM}{BEM object}
}
\description{
ADD LATER
}
