% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\docType{data}
\name{temps_casc}
\alias{temps_casc}
\title{Daily water temperature for many streams}
\format{
A data frame with 365 rows and 515 columns:
\describe{
  \item{date}{(date) dates from 1 January 2022 to 31 December 2022.}
  \item{COMID}{(numeric) mean daily water temperature in degrees C for all 513 COMIDs.}
  ...
}
}
\source{
Southeastern Climate Adaptation Science Center (CASC).
}
\usage{
temps_casc
}
\description{
This dataset contains modeled maximum daily water temperature for every day of the year 2022 for all interconfluence reaches of the Little River, TN.
}
\keyword{datasets}
