% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishyenergy.R
\docType{data}
\name{temps_nwis}
\alias{temps_nwis}
\title{Daily water temperature for one stream}
\format{
A data frame with 365 rows and 2 columns:
\describe{
\item{date}{(date) dates from 1 January 2025 to 31 December 2025}
\item{WT_mean}{(numeric) mean daily water temperature in degrees C}
...
}
}
\source{
National Water Information System.
}
\usage{
temps_nwis
}
\description{
This dataset contains mean daily water temperature for every day of the year 2025 from the USGS gage on the Little River above Townsend, TN (03497300).
}
\keyword{datasets}
